player/playloop.c: Revert --loop-file and --start interaction
This reverts commit c247d8ecab3b8a255510dd4c1522dcbdd0fad49f
and commit c52e15f2f8e6947c583638d32c4ee79263feb0c4.

I had changed --loop-file to interact with --start to work
the same way that --loop-playlist does. (That is, --loop-file
seeks to the --start time upon looping, not the beginning of
the file.) However, the consensus is that the old behavior is
preferred and the interaction with --loop-playlist is the one
that is incorrect.

In addition, this change introduced a bug in the interaction
between Quit-Watch-Later and --loop-file, where upon reaching
playback end it would seek to the resume timestamp, not the
start of the file.

As a result, this commit reverts that change.