Revert "command: make pause display the same osd-msg-bar as seek"
This reverts commit 9812e276aa1bb0bddeb73677aa9e9f87e73cd930.

This was apparently unpopular. I still think the pause OSD should be the
same as seek even if it's not visible by default, but it seems that
whether to display a given property change is currently conflated with
what to display.

The reverted behaviour can be restored by adding something like the
following to input.conf:

SPACE cycle pause; show_progress